Refactor utils into own file
