+++ packages/grafana-ui/src/components/Icon/Icon.tsx
@@ -6,8 +6,7 @@ import { useTheme } from '../../themes/ThemeContext';
 import { IconName, IconType, IconSize } from '../../types/icon';
 import SVG from 'react-inlinesvg';
 import { cacheInitialized, initIconCache, iconRoot } from './iconBundle';
-
-const alwaysMonoIcons: IconName[] = ['grafana', 'favorite', 'heart-break', 'heart', 'panel-add', 'library-panel'];
+import { getIconSubDir, getSvgSize } from './utils';
 
 export interface IconProps extends React.HTMLAttributes<HTMLDivElement> {
   name: IconName;
@@ -33,20 +32,6 @@ const getIconStyles = stylesFactory((theme: GrafanaTheme) => {
     `,
   };
 });
-
-export function getIconSubDir(name: IconName, type: string): string {
-  if (name?.startsWith('gf-')) {
-    return 'custom';
-  } else if (alwaysMonoIcons.includes(name)) {
-    return 'mono';
-  } else if (type === 'default') {
-    return 'unicons';
-  } else if (type === 'solid') {
-    return 'solid';
-  } else {
-    return 'mono';
-  }
-}
 
 export const Icon = React.forwardRef<HTMLDivElement, IconProps>(
   ({ size = 'md', type = 'default', name, className, style, title = '', ...divElementProps }, ref) => {
@@ -98,23 +83,3 @@ function getFontAwesomeIconStyles(iconName: string, className?: string): string 
     className
   );
 }
-
-/* Transform string with px to number and add 2 pxs as path in svg is 2px smaller */
-export const getSvgSize = (size: IconSize) => {
-  switch (size) {
-    case 'xs':
-      return 12;
-    case 'sm':
-      return 14;
-    case 'md':
-      return 16;
-    case 'lg':
-      return 18;
-    case 'xl':
-      return 24;
-    case 'xxl':
-      return 36;
-    case 'xxxl':
-      return 48;
-  }
-};
