+++ packages/grafana-ui/src/components/Icon/utils.test.ts
@@ -1,4 +1,4 @@
-import { getIconSubDir } from './Icon';
+import { getIconSubDir } from './utils';
 
 describe('Icon', () => {
   describe('getIconSubDir', () => {
