+++ packages/grafana-ui/src/components/Icon/utils.ts
@@ -1,0 +1,37 @@
+import { IconName, IconSize } from '../../types/icon';
+
+const alwaysMonoIcons: IconName[] = ['grafana', 'favorite', 'heart-break', 'heart', 'panel-add', 'library-panel'];
+
+export function getIconSubDir(name: IconName, type: string): string {
+  if (name?.startsWith('gf-')) {
+    return 'custom';
+  } else if (alwaysMonoIcons.includes(name)) {
+    return 'mono';
+  } else if (type === 'default') {
+    return 'unicons';
+  } else if (type === 'solid') {
+    return 'solid';
+  } else {
+    return 'mono';
+  }
+}
+
+/* Transform string with px to number and add 2 pxs as path in svg is 2px smaller */
+export const getSvgSize = (size: IconSize) => {
+  switch (size) {
+    case 'xs':
+      return 12;
+    case 'sm':
+      return 14;
+    case 'md':
+      return 16;
+    case 'lg':
+      return 18;
+    case 'xl':
+      return 24;
+    case 'xxl':
+      return 36;
+    case 'xxxl':
+      return 48;
+  }
+};
